Add gene frequency

Add gene-level frequency, i.e. how often mutations in a given gene are associated with a given phenotype. Numeric frequency columns are on a 0-100% scale.

add_gene_frequency(
  phenotype_to_genes = load_phenotype_to_genes(),
  all.x = TRUE,
  verbose = TRUE
)

Arguments

phenotype_to_genes

Output of load_phenotype_to_genes mapping phenotypes to gene annotations.

all.x

logical; if TRUE, then extra rows will be added to the output, one for each row in x that has no matching row in y. These rows will have 'NA's in those columns that are usually filled with values from y. The default is FALSE, so that only rows with data from both x and y are included in the output.

verbose

Print messages.

Value

phenos data.table with extra column

Examples

phenotype_to_genes <- load_phenotype_to_genes()[seq_len(1000),]
#> Importing existing file: ... phenotype_to_genes.txt
phenos2 <- add_gene_frequency(phenotype_to_genes = phenotype_to_genes)
#> Adding HPO IDs.
#> Translating all phenotypes to HPO IDs.
#> + Returning a vector of phenotypes (same order as input).
#> Annotating gene frequencies.
#> Importing existing file: ... genes_to_phenotype.txt