Get term definition

This function accesses the HPO API to get a description/definition of an HPO term. If a line_length \> 0 is passed to the function, it will add newlines every nth word. This can be useful when displaying the description in plots with limited space.

add_hpo_definition(
  phenos,
  line_length = FALSE,
  use_api = FALSE,
  verbose = TRUE
)

Arguments

phenos

dataframe of phenotypes and values / parameters.

line_length

The number of desired words per line \<int\>

use_api

Get definitions from the HPO API, as opposed to a static local dataset.

verbose

Print messages.

Value

A named vector of HPO term descriptions.

Examples

phenos <- example_phenos()
#> ℹ All local files already up-to-date!
phenos2 <- add_hpo_definition(phenos = phenos)
#> Adding term definitions.
#> ℹ All local files already up-to-date!