Add HPO ID column to dataframe

This adds the HPO term id column to the subset of ewce results data to be plotted in the cell select app. It also checks if it is a valid HPO term id to prevent error and adds a boolean column where TRUE if term is valid. If the HPO Id is not correct, it caused an error in the ontologyPlot package.

add_hpo_id(phenos, hpo = get_hpo(), phenotype_to_genes = NULL, verbose = FALSE)

Arguments

phenos

dataframe of phenotypes and values / parameters.

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

phenotype_to_genes

Output of load_phenotype_to_genes mapping phenotypes to gene annotations.

verbose

Print messages.

Value

The subset of EWCE results dataframe with a HPO Id column added.

Examples

phenotype_to_genes <- load_phenotype_to_genes()
#> Reading cached RDS file: phenotype_to_genes.txt
#> + Version: v2023-10-09
phenos <- unique(phenotype_to_genes[,c("hpo_id","hpo_name")])
phenos2 <- add_hpo_id(phenos=phenos)