Add age of death

Add metadata from MONDO, including:

MONDO_ID: MONDO term ID.

MONDO_name: MONDO term name.

MONDO_definition: MONDO term definition.

add_mondo(
  phenos,
  add_definitions = TRUE,
  all.x = TRUE,
  allow.cartesian = FALSE,
  verbose = TRUE
)

Arguments

phenos

dataframe of phenotypes and values / parameters.

add_definitions

Add disease definitions using add_disease_definition.

all.x

logical; if TRUE, then extra rows will be added to the output, one for each row in x that has no matching row in y. These rows will have 'NA's in those columns that are usually filled with values from y. The default is FALSE, so that only rows with data from both x and y are included in the output.

allow.cartesian

See allow.cartesian in [.data.table.

verbose

Print messages.

Examples

phenos <- load_phenotype_to_genes("phenotype.hpoa")
phenos2 <- add_mondo(phenos = phenos)
#> Adding disease metadata: Definitions, Preferred.Label
#> Importing Orphanet metadata.
#> Importing OMIM metadata.
#> 0 / 12527 (0%) disease_name missing.
#> 8228 / 12468 (65.99%) Definitions missing.
#> Annotating phenos with MONDO metadata.
#> ℹ All local files already up-to-date!
#> 129 / 12468 (1.03%) MONDO_ID missing.
#> 11116 / 12468 (89.16%) MONDO_name missing.
#> 12266 / 12468 (98.38%) MONDO_definition missing.
#> 8100 / 12468 (64.97%) Definitions missing.
#> Annotating phenos with MONDO metadata.
#> ℹ All local files already up-to-date!
#> 129 / 12468 (1.03%) MONDO_ID missing.
#> 11116 / 12468 (89.16%) MONDO_name missing.
#> 12266 / 12468 (98.38%) MONDO_definition missing.