Add age of onset

Add age of onset for each HPO ID. Onset IDs and assigned "Onset_score" values:

HP:0011461 "Fetal onset" (Onset_score=1)

HP:0030674 "Antenatal onset" (Onset_score=2)

HP:0003577 "Congenital onset" (Onset_score=3)

HP:0003623 "Neonatal onset" (Onset_score=4)

HP:0003593 "Infantile onset" (Onset_score=5)

HP:0011463 "Childhood onset" (Onset_score=6)

HP:0003621 "Juvenile onset" (Onset_score=7)

HP:0011462 "Young adult onset" (Onset_score=8)

HP:0003581 "Adult onset" (Onset_score=9)

HP:0003596 "Middle age onset" (Onset_score=10)

HP:0003584 "Late onset" (Onset_score=11)

add_onset(
  phenos,
  keep_onsets = NULL,
  agg_by = NULL,
  all.x = TRUE,
  allow.cartesian = FALSE,
  verbose = TRUE
)

Arguments

phenos

dataframe of phenotypes and values / parameters.

keep_onsets

The age of onset associated with each HPO ID to keep. If >1 age of onset is associated with the term, only the earliest age is considered. See add_onset for details.

agg_by

Column to aggregate age of onset metadata by.

all.x

logical; if TRUE, then extra rows will be added to the output, one for each row in x that has no matching row in y. These rows will have 'NA's in those columns that are usually filled with values from y. The default is FALSE, so that only rows with data from both x and y are included in the output.

allow.cartesian

See allow.cartesian in [.data.table.

verbose

Print messages.

Value

phenos data.table with extra columns:

"Onset": Onset HPO IDs of disease phenotypes associated with the target HPO_ID phenotype.

"Onset_names": Onset HPO names of disease phenotypes associated with the target HPO_ID phenotype.

"Onset_counts": The number of times each term in "Onset_names" appears across associated disease phenotypes.

"Onset_score_mean": Mean onset score.

"Onset_score_min": Minimum onset score.

"Onset_top": The most common onset term.

"Onset_earliest": The earliest age of onset.

"Onset_latest": The latest age of onset.

Examples

phenos <- example_phenos()
phenos2 <- add_onset(phenos = phenos)
#> Annotating phenos with Onset.
#> Annotating phenos with Disease
#> Importing existing file: ... phenotype.hpoa