Add severity Tiers (auto)

Automatically add severity Tier for each HPO ID, in accordance with the rating system provided by Lazarin et al (2014). In order of increasing severity:

Tier 4 Reduced fertility

Tier 3 Sensory impairment: vision, Immunodeficiency/cancer, Sensory impairment: hearing, Sensory impairment: touch, other (including pain), Mental illness, Dysmorphic features

Tier 2 Shortened life span: premature adulthood, Impaired mobility, Internal physical malformation

Tier 1 Shortened life span: infancy, Shortened life span: childhood/adolescence, Intellectual disability

assign_tiers(
  hpo = get_hpo(),
  terms = hpo$id,
  keyword_sets = list(Tier1 = c("intellectual disability", "death"), Tier2 =
    c("impaired mobility", "malform"), Tier3 = c("sight", "vision", "immunodeficien",
    "cancer", "hearing", "touch", "pain ", " pain", "mental illness", "dysmorphic"),
    Tier4 = c("fertility")),
  check_names = TRUE,
  check_definitions = TRUE,
  check_diseases = FALSE,
  search_ancestors = TRUE,
  search_descendants = TRUE,
  as_datatable = FALSE,
  verbose = TRUE
)

Arguments

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

terms

A subset of HPO IDs to assign Tiers to.

keyword_sets

A named list of regex queries to be used when searching for phenotypes that have matching metadata.

check_names

When regexsearching for matching terms, check the phenotype names.

check_definitions

When regex searching for matching terms, check the phenotype definition.

check_diseases

When regex searching for matching terms, check the names of diseases associated with the phenotype.

search_ancestors

Inherit Tiers of ancestors.

search_descendants

Inherit Tiers of descendants.

as_datatable

Return the results as a data.table.

verbose

Print messages.

Value

Tier assignments for each term in terms. Will be returned as either a named vector or a data.table.

Examples

terms <- get_hpo()$id[seq(10)]
#> ℹ All local files already up-to-date!
tiers <- assign_tiers(terms = terms)
#> ℹ All local files already up-to-date!
#> keyword_set=Tier1: Found 58 terms matching category.
#> keyword_set=Tier2: Found 135 terms matching category.
#> keyword_set=Tier3: Found 400 terms matching category.
#> keyword_set=Tier4: Found 8 terms matching category.
#> Auto-assigning severity Tiers to 10 terms.