Get functions

Functions to get data objects and extract elements.

get_hpo(
  lvl = 2,
  force_new = FALSE,
  terms = NULL,
  method = "github",
  save_dir = KGExplorer::cache_dir(package = "HPOExplorer"),
  ...
)

get_hpo_id_direct(term, hpo = get_hpo())

get_mappings(
  terms = NULL,
  to = c("UMLS", "SNOMED", "ICD9", "ICD10"),
  max_dist = 1,
  force_new = FALSE,
  save_dir = KGExplorer::cache_dir(package = "HPOExplorer"),
  ...
)

Arguments

lvl

How many levels deep into the ontology to get ancestors from. For example:

1: "All"

2: "Phenotypic abnormality"

3: "Abnormality of the nervous system"

4: "Abnormality of nervous system physiology"

5: "Neurodevelopmental abnormality" or "Behavioral abnormality"

force_new

Force a new download or creation of a data resource.

terms

A subset of HPO IDs to assign Tiers to.

method

Method to construct plot with.

save_dir

Directory to save a file to.

...

Arguments passed on to KGExplorer::get_ontology, get_data

name

<...>Any ontology name from get_ols_options

"mondo" Import the Mondo ontology. GitHub..

"hpo" Import the Human Phenotype Ontology. GitHub.

filetype

File type to search for.

import_func

Function to import the ontology with. If NULL, automatically tries to choose the correct function.

add_metadata

Add metadata to the resulting ontology object.

add_n_edges

Add the number of edges (connections) for each term.

add_ontology_levels

Add the ontology level for each term.

add_version

Add the release version to the returned object's attributes

file

name or vector of names of files to be downloaded. If NULL, all assets attached to the release will be downloaded.

repo

Repository name in format "owner/repo". Defaults to guess_repo().

tag

tag for the GitHub release to which this data should be attached.

overwrite

Should any local files of the same name be overwritten? default TRUE.

term

One or more ontology IDs.

hpo

Human Phenotype Ontology object, loaded from get_ontology.

to

Medical ontologies to provide mappings to.

max_dist

Maximum cross-ontology distance to search for mappings. A distance of 1 means only direct mappings are returned. Greater distances mean that mappings are found through intermediate terms and are therefore less reliable.

Value

Data.

ontology_DAG object.

The HPO ID(s) of phenotype(s)

A named list of data.table objects containing mappings.

Functions

get_hpo(): get_ Get Human Phenotype Ontology (HPO)

Updated version of Human Phenotype Ontology (HPO). Created from the OBO files distributed by the HPO project's GitHub.

By comparison, the hpo data from ontologyIndex is from 2016. Note that the maximum ontology level depth in the 2016 version was 14, whereas in the 2023 version the maximum ontology level depth is 16 (due to an expansion of the HPO).

get_hpo_id_direct(): get_ Get HPO term ID direct

Directly retrieves it from the HPO ontology object.

get_mappings(): get_ Get Human Phenotype Ontology (HPO) mappings

Get mappings between HPO terms and their corresponding IDs in other medical ontologies (e.g. SNOMED CT, UMLS, ICD-9/10).

Examples

hpo <- get_hpo()
term = "Phenotypic abnormality"
pheno_abnormality_id <- get_hpo_id_direct(term = term)
mappings <- get_mappings()