Get remote data

Download remotely stored data via pb_download.

get_data(
  file,
  tag = "latest",
  repo = "neurogenomics/HPOExplorer",
  save_dir = KGExplorer::cache_dir(package = "HPOExplorer"),
  add_version = FALSE,
  overwrite = TRUE
)

Arguments

file

name or vector of names of files to be downloaded. If NULL, all assets attached to the release will be downloaded.

tag

tag for the GitHub release to which this data should be attached.

repo

Repository name in format "owner/repo". Defaults to guess_repo().

save_dir

Directory to save data to.

add_version

Add the release version to the returned object's attributes

overwrite

Should any local files of the same name be overwritten? default TRUE.

Value

Path to downloaded file or the object itself (when ".rds" format).