Get GenCC

Get phenotype-gene evidence score from the Gene Curation Coalition. Data downloaded from here.
NOTE: Due to licensing restrictions, a GenCC download does not include OMIM data. OMIM data can be accessed and downloaded through OMIM.
NOTE: GenCC does not currently have any systematic versioning. There for the attr(obj,"version") attribute is set to the date it was downloaded and cached by get_gencc.

get_gencc(
  agg_by = c("disease_id", "gene_symbol"),
  dict = c(Definitive = 6, Strong = 5, Moderate = 4, Supportive = 3, Limited = 2,
    `Disputed Evidence` = 1, `Refuted Evidence` = 0, `No Known Disease Relationship` = 0),
  save_dir = tools::R_user_dir("HPOExplorer", which = "cache"),
  verbose = TRUE
)

Arguments

agg_by

Column to aggregate metadata by.

dict

A named vector of evidence score mappings. See here for more information.

save_dir

Directory to save data to.

verbose

Print messages.

Value

data.table with columns:

"disease_id": Disease ID.

"gene_symbol": Gene symbol.

"evidence_score": Evidence score.

Examples

d <- get_gencc()
#> Gathering data from GenCC.
#> Importing cached file.
#> + Version: 2023-11-15