Get Human Phenotype Ontology (HPO)

Updated version of Human Phenotype Ontology (HPO) from 2023-10-09. Created from the OBO files distributed by the HPO project's GitHub.

get_hpo(
  save_dir = tools::R_user_dir("HPOExplorer", which = "cache"),
  tag = "latest",
  overwrite = TRUE
)

Source

BioPortal f <- HPOExplorer:::get_data(file = "hp-base.obo", repo = "obophenotype/human-phenotype-ontology", tag="latest") hpo <- ontologyIndex::get_OBO(f, extract_tags = "everything") ### Fix non-ASCII characters in metadata #### func <- function(v){ Encoding(v) <- "latin1" iconv(v, "latin1", "UTF-8") } attributes(hpo)$version <- func(attributes(hpo)$version) save_path <- file.path(tempdir(),"hpo.rds") saveRDS(hpo,save_path) piggyback::pb_upload(file = save_path, repo = "neurogenomics/HPOExplorer", overwrite = TRUE, tag = "latest")

Arguments

save_dir

Directory to save data to.

tag

tag for the GitHub release to which this data should be attached.

overwrite

Should any local files of the same name be overwritten? default TRUE.

Value

ontology_index

Details

By comparison, the hpo data from ontologyIndex is from 2016. Note that the maximum ontology level depth in the 2016 version was 14, whereas in the 2023 version the maximum ontology level depth is 16 (due to an expansion of the HPO).

Examples

hpo <- get_hpo()
#> ℹ All local files already up-to-date!