Generate network plot

This creates a network plot which is compatible with plotly to make interactive plots. It makes it possible to hover box that includes your results related to each phenotype and/or a description of the phenotypes.

ggnetwork_plot(
  phenoNet,
  colour_var = "fold_change",
  colour_label = gsub("_", " ", colour_var),
  size_var = "ontLvl",
  interactive = TRUE,
  tooltip = "hover",
  verbose = TRUE,
  ...
)

Arguments

phenoNet

The network object created using create_network_object

colour_var

Column to be mapped to node colour.

colour_label

A label for the colour figure legend \<string\>.

size_var

Column name to be mapped to node size.

interactive

Make the plot interactive with ggplotly.

tooltip

a character vector specifying which aesthetic mappings to show in the tooltip. The default, "all", means show all the aesthetic mappings (including the unofficial "text" aesthetic). The order of variables here will also control the order they appear. For example, use tooltip = c("y", "x", "colour") if you want y first, x second, and colour last.

verbose

Print messages.

...

Arguments passed on to plotly::ggplotly

p

a ggplot object.

width

Width of the plot in pixels (optional, defaults to automatic sizing).

height

Height of the plot in pixels (optional, defaults to automatic sizing).

dynamicTicks

should plotly.js dynamically generate axis tick labels? Dynamic ticks are useful for updating ticks in response to zoom/pan interactions; however, they can not always reproduce labels as they would appear in the static ggplot2 image.

layerData

data from which layer should be returned?

originalData

should the "original" or "scaled" data be returned?

source

a character string of length 1. Match the value of this string with the source argument in event_data() to retrieve the event data corresponding to a specific plot (shiny apps can have multiple plots).

Value

A network plot (compatible with interactive plotly rendering).

Examples

phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
#> ℹ All local files already up-to-date!
#> Reading cached RDS file: phenotype_to_genes.txt
#> + Version: v2023-10-09
#> Extracting data for 23 descendents.
#> Computing gene counts.
#> Getting absolute ontology level for 23 HPO IDs.
#> Computing ontology level / gene count ratio.
#> Adding term definitions.
#> ℹ All local files already up-to-date!
#> Adding level-3 ancestor to each HPO ID.
#> ℹ All local files already up-to-date!
phenoNet <- make_network_object(phenos = phenos,
                                colour_var = "ontLvl_geneCount_ratio")
#> Making phenotype network object.
#> Creating adjacency matrix for 23 term(s).
#> ℹ All local files already up-to-date!
#> Creating ggnetwork object.
#> Adding new_column='hpo_name'
#> Adding new_column='hpo_id'
#> Adding new_column='ontLvl'
#> Adding new_column='ontLvl_geneCount_ratio'
#> Adding new_column='definition'
#> Adding new_column='ancestor'
#> Adding new_column='ancestor_name'
#> Adding n_edges per node.
plt <- ggnetwork_plot(phenoNet = phenoNet,
                      colour_var = "ontLvl_geneCount_ratio",
                      colour_label = "ontLvl_genes")
#> Creating ggnetwork plot.