Check annotations from GPT

Check GPT phenotype annotations using a several metrics.

gpt_annot_check(
  annot = gpt_annot_read(),
  query_hits = search_hpo(),
  response_map = list(no = "never", yes = "always"),
  pos_values = c("sometimes", "often", "always"),
  neg_values = c("never", "rarely"),
  verbose = TRUE
)

Arguments

annot

GPT-generated phenotype annotations.

query_hits

A named list of HPO IDs for each query, generated by search_hpo. These will be used as the ground truth when trying to identify true positive annotations.

response_map

A named list of responses to map onto. Used for standardisation of responses.

pos_values

Positive values.

neg_values

Negative values.

verbose

Print messages.

Value

Named list

Examples

checks <- gpt_annot_check()
#> Translating ontology terms to ids.
#> Reading cached RDS file: phenotype_to_genes.txt
#> + Version: 2025-03-03
#> 151 phenotypes do not have matching HPO IDs.
#> Reading in GPT annotations for 16,982 phenotypes.
#> Querying HPO for matching terms.
#> Number of phenotype hits per query group:
#>  - intellectual_disability: 19
#>  - impaired_mobility: 319
#>  - physical_malformations: 78
#>  - blindness: 1
#>  - sensory_impairments: 252
#>  - immunodeficiency: 29
#>  - cancer: 56
#>  - reduced_fertility: 9
#> Warning: The `facets` argument of `facet_grid()` is deprecated as of ggplot2 2.2.0.
#> ℹ Please use the `rows` argument instead.
#> ℹ The deprecated feature was likely used in the HPOExplorer package.
#>   Please report the issue at
#>   <https://github.com/neurogenomics/HPOExplorer/issues>.