Harmonise phenotypes

Harmonise a mixed vector of phenotype names (e.g. "Focal motor seizure") and HPO IDs (e.g. c("HP:0000002","HP:0000003")).

harmonise_phenotypes(
  phenotypes,
  hpo = get_hpo(),
  as_hpo_ids = FALSE,
  validate = TRUE,
  ignore_case = TRUE,
  keep_order = TRUE,
  invert = FALSE,
  verbose = TRUE
)

Arguments

phenotypes

A character vector of phenotype names and/or HPO IDs.

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

as_hpo_ids

Return all phenotypes as HPO IDs (TRUE), or phenotype names (FALSE).

validate

Remove any phenotype names or IDs that cannot be found in the hpo.

ignore_case

Ignore casing when searching for matches.

keep_order

Return a named list of the same length and order as phenotypes. If FALSE, return a named list of only the unique phenotypes, sometimes in a different order.

invert

Invert the keys/values of the dictionary, such that the key becomes the values (and vice versa).

verbose

Print messages.

Value

Character vector

Examples

phenotypes <- c("Focal motor seizure","HP:0000002","HP:0000003")
#### As phenotype names ####
pheno_names <- harmonise_phenotypes(phenotypes=phenotypes)
#> Translating all phenotypes to names.
#> ℹ All local files already up-to-date!
#> + Returning a vector of phenotypes (same order as input).
#### As HPO IDs ####
pheno_ids <- harmonise_phenotypes(phenotypes=phenotypes, as_hpo_ids=TRUE)
#> Translating all phenotypes to HPO IDs.
#> ℹ All local files already up-to-date!
#> + Returning a vector of phenotypes (same order as input).