Make Hoverbox

A hoverbox is a box of text that shows up when the cursor hovers over something. These can be useful when making interactive network plots of the HPO phenotypes because we can include a hoverbox that gives information and data associated with each phenotype.

make_hoverbox(
  phenotype,
  phenos_dataframe,
  columns = c("HPO_Id", "description"),
  labels = c("ID", "Description")
)

Arguments

phenotype

The phenotype you are making a hoverbox for

phenos_dataframe

The dataframe of phenotypes and parameters to be included

columns

chr vector of column names from the phenos_dataframe

labels

chr vector of labels for the selected columns

Value

A nicely formatted string with newlines etc, to be used as a hoverbox

Details

This function expects a dataframe of with a "Phenotype" column that has the name of each phenotype. It must then include columns for all of the parameters you wish to include in the hoverbox.