Make the Human Phenotype Ontology

Make the Human Phenotype Ontology (HPO) ontology_index object from the Open Biomedical Ontologies (OBO) file distributed via the official HPO GitHub Releases.

make_hpo(
  file = "hp-base.obo",
  repo = "obophenotype/human-phenotype-ontology",
  tag = "latest",
  save_dir = tools::R_user_dir(package = "HPOExplorer", which = "cache"),
  extract_tags = "everything",
  fix_ascii = TRUE,
  overwrite = TRUE,
  upload_tag = NULL,
  ...
)

Arguments

file

name or vector of names of files to be downloaded. If NULL, all assets attached to the release will be downloaded.

repo

Repository name in format "owner/repo". Defaults to guess_repo().

tag

tag for the GitHub release to which this data should be attached.

save_dir

Directory to save data to.

extract_tags

Character value: either "minimal" or "everything", determining whether to extract only the properties of terms which are required to run functions in the package - i.e. "id", "name", "parents", "children" and "ancestors" - or extract all properties provided in the file. Term properties are named in the resulting ontology_index as their corresponding tags in the OBO file (except "parents", "children" and "ancestors" which are appended with "_OBO" to avoid clashing with standard ontology_index properties. Defaults to "minimal".

fix_ascii

Fix non-ASCII characters in metadata.

overwrite

Should any local files of the same name be overwritten? default TRUE.

...

Arguments passed on to ontologyIndex::get_OBO

propagate_relationships

Character vector of relations

merge_equivalent_terms

Logical value determining whether terms that are marked "equivalent_to" a target term should be merged, retaining properties of the target term when the property should have one value, e.g. the term ID and name. Defaults to TRUE.

Value

ontology_index object