Make phenotypes dataframe

Make a dataframe from a subset of the Human Phenotype Ontology.

make_phenos_dataframe(
  ancestor,
  hpo = get_hpo(),
  phenotype_to_genes = load_phenotype_to_genes(),
  add_description = TRUE,
  add_hoverboxes = TRUE,
  add_age_onset = FALSE,
  add_severity_tiers = FALSE,
  columns = list(Phenotype = "Phenotype", ID = "HPO_ID", ontLvl = "ontLvl", ontLvl_genes
    = "ontLvl_geneCount_ratio", Definition = "definition"),
  interactive = TRUE,
  verbose = TRUE
)

Arguments

ancestor

The ancestor to get all descendants of. If NULL, returns the entirely ontology.

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

phenotype_to_genes

Output of load_phenotype_to_genes mapping phenotypes to gene annotations.

add_description

Whether to get the phenotype descriptions as well (slower).

add_hoverboxes

Add hoverdata with make_hoverboxes.

add_age_onset

Add age of onset columns using add_onset.

add_severity_tiers

Add severity Tiers column using add_tier.

columns

A named vector of columns in phenos to add to the hoverdata via make_hoverboxes.

interactive

Make the plot interactive with ggplotly.

verbose

Print messages.

Value

The HPO in dataframe format.

Examples

phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
#> Importing existing file: /github/home/.cache/R/HPOExplorer/data/phenotype_to_genes.txt
#> Extracting data for 17 descendents.
#> Computing gene counts.
#> Getting ontology level for 17 HPO IDs.
#> Creating adjacency matrix for 17 terms.
#> Computing ontology level / gene count ratio
#> Adding information_content scores.
#> Adding term definitions.
#> Making hoverboxes from: 'Phenotype', 'HPO_ID', 'ontLvl', 'ontLvl_geneCount_ratio', 'definition'