Search the Human Phenotype Ontology

Serarch for phenotypes in the HPO that match substring queries, or are the descendants of those substring-matched phenotypes.

search_hpo(
  hpo = get_hpo(),
  queries = list(intellectual_disability = c("Intellectual disability",
    "Mental deterioration"), impaired_mobility = c("Gait disturbance",
    "Diminished movement", " mobility"), physical_malformations = c("malformation"),
    blindness = c("^blindness"), sensory_impairments = c("Abnormality of vision",
    "Abnormality of the sense of smell", "Abnormality of taste sensation",
    "Somatic sensory dysfunction", "Hearing abnormality"), immunodeficiency =
    c("Immunodeficiency", "Impaired antigen-specific response"), cancer = c("Cancer", 
  
      "malignant", "carcinoma"), reduced_fertility = c("Decreased fertility",
    "Hypogonadism")),
  search_cols = c("name"),
  include_descendants = TRUE,
  include_ancestors = FALSE,
  ignore.case = TRUE,
  fixed = FALSE,
  verbose = TRUE
)

Arguments

hpo

Human Phenotype Ontology object, loaded from get_ontology.

queries

A named list of character vectors. Each vector includes a phenoptype name (or a substring) to be searched for in the HPO.

search_cols

HPO metadata column names to query for hits.

include_descendants

Include all descendants of query hits.

include_ancestors

Include all ancestors of query hits.

ignore.case

if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.

fixed

logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.

verbose

Print messages.

Value

Named list of HPO IDs.

Examples

query_hits <- search_hpo()
#> Querying HPO for matching terms.
#> Number of phenotype hits per query group:
#>  - intellectual_disability: 19
#>  - impaired_mobility: 326
#>  - physical_malformations: 81
#>  - blindness: 1
#>  - sensory_impairments: 253
#>  - immunodeficiency: 29
#>  - cancer: 56
#>  - reduced_fertility: 9