Subset descendants

Subset a phenos datatable to only descendants of an ancestor HPO ID term.

subset_descendants(
  phenos,
  ancestor = NULL,
  hpo = get_hpo(),
  ignore_case = TRUE,
  verbose = TRUE
)

Arguments

phenos

dataframe of phenotypes and values / parameters.

ancestor

Phenotype names (or HPO ID) of an ancestor in the Human Phenotype Ontology. Only phenotypes that are descendants of this ancestor will be kept. Set to NULL (default) to skip this filtering step.

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

ignore_case

Ignore casing when searching for matches.

verbose

Print messages.

Value

data.table of phenotypes, with additional columns: "ancestor", "ancestor_id"

Examples

phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
#> ℹ All local files already up-to-date!
#> Reading cached RDS file: phenotype_to_genes.txt
#> + Version: v2023-10-09
#> Extracting data for 23 descendents.
#> Computing gene counts.
#> Getting absolute ontology level for 23 HPO IDs.
#> Computing ontology level / gene count ratio.
#> Adding term definitions.
#> ℹ All local files already up-to-date!
#> Adding level-3 ancestor to each HPO ID.
#> ℹ All local files already up-to-date!
phenos2 <- subset_descendants(phenos = phenos,
                              ancestor = "Motor delay")
#> Subsetting phenotypes to only ancestors of: Motor delay
#> Translating all phenotypes to HPO IDs.
#> ℹ All local files already up-to-date!
#> + Returning a vector of phenotypes (same order as input).
#> 9 associations remain after filtering.