Generate network plot

This creates a network plot which is compatible with plotly to make interactive plots. It makes it possible to hover box that includes your results related to each phenotype and/or a description of the phenotypes.

ggnetwork_plot(
  phenoNet,
  colour_column = "fold_change",
  colour_label = "Fold change",
  interactive = FALSE,
  tooltip = "hover",
  verbose = TRUE
)

Arguments

phenoNet

The network object created using create_network_object

colour_column

column name of the variable to be mapped to colour \<string\>

colour_label

A label for the colour figure legend \<string\>.

interactive

Make the plot interactive with ggplotly.

tooltip

a character vector specifying which aesthetic mappings to show in the tooltip. The default, "all", means show all the aesthetic mappings (including the unofficial "text" aesthetic). The order of variables here will also control the order they appear. For example, use tooltip = c("y", "x", "colour") if you want y first, x second, and colour last.

verbose

Print messages.

Value

A network plot (compatible with interactive plotly rendering).

Examples

library(ontologyIndex)
data(hpo)
phenos = make_phenos_dataframe(hpo = hpo,
                               ancestor = "Neurodevelopmental delay",
                               add_description = FALSE)
#> Importing existing file: /github/home/.cache/R/HPOExplorer/data/phenotype_to_genes.txt
#> Extracting data for 17 descendents.
#> Processing: Delayed speech and language development
#> Processing: Global developmental delay
#> Processing: Motor delay
#> Processing: Absent speech
#> Processing: Delayed gross motor development
#> Processing: Expressive language delay
#> Processing: Severe expressive language delay
#> Processing: Delayed fine motor development
#> Processing: Receptive language delay
#> Processing: Mild global developmental delay
#> Processing: Moderate global developmental delay
#> Processing: Severe global developmental delay
#> Processing: Moderate receptive language delay
#> Processing: Severe receptive language delay
#> Processing: Delayed social development
#> Processing: Profound global developmental delay
#> Processing: Neurodevelopmental delay
phenos <- make_hoverboxes(phenos_dataframe = phenos)
adjacency <- adjacency_matrix(unique(phenos$HPO_Id), hpo)
#> adjacency_matrix
phenoNet <- make_network_object(phenos = phenos,
                                adjacency = adjacency,
                                hpo = hpo,
                                colour_column = "ontLvl_geneCount_ratio")
plt <- ggnetwork_plot(phenoNet = phenoNet,
                      colour_column = "ontLvl_geneCount_ratio",
                      colour_label = "OntLvl/nGenes")
#> Creating ggnetwork plot.