Phenotypes to GenomicRanges

Convert a HPO phenotype dataframe generated by make_phenos_dataframe to a GRangesList split by HPO ID. The resulting object will contain genes (and gene metadata) for all genes associated with each phenotypes.

phenos_to_granges(
  phenos,
  phenotype_to_genes = load_phenotype_to_genes(),
  hpo = get_hpo(),
  keep_seqnames = c(seq_len(22), "X", "Y"),
  split.field = "ID",
  verbose = TRUE
)

Arguments

phenos

dataframe of phenotypes and values / parameters.

phenotype_to_genes

Output of load_phenotype_to_genes mapping phenotypes to gene annotations.

hpo

Human Phenotype Ontology object, loaded from ontologyIndex.

keep_seqnames

Chromosomes to keep.

split.field

A character string of a recognized column name in df that contains the grouping. This column defines how the rows of df are split and is typically a factor or character vector. When split.field is not provided the df will be split by the number of rows.

verbose

Print messages.

Value

A GRangesList.

Examples

phenos <- make_phenos_dataframe(ancestor = "Neurodevelopmental delay")
#> Importing existing file: ... phenotype_to_genes.txt
#> Extracting data for 23 descendents.
#> Computing gene counts.
#> Getting absolute ontology level for 23 HPO IDs.
#> Computing ontology level / gene count ratio.
#> Adding information_content scores.
#> Adding term definitions.
#> Adding level-3 ancestor to each HPO ID.
#> Making hoverboxes from: 'Phenotype', 'HPO_ID', 'ontLvl', 'ontLvl_geneCount_ratio', 'definition', 'ancestor', 'ancestor_name'
grl <- phenos_to_granges(phenos = phenos)
#> Converting phenos to GRangesList.
#> Importing existing file: ... phenotype_to_genes.txt
#> Translating all phenotypes to HPO IDs.
#> + Returning a vector of phenotypes (same order as input).
#> Translating all phenotypes to HPO IDs.
#> + Returning a vector of phenotypes (same order as input).
#> Loading required namespace: ensembldb
#> Gathering gene metadata
#> Loading required namespace: EnsDb.Hsapiens.v75